refactor: modularize overview_render.c into separate functional components
Decomposed the monolithic overview_render.c into targeted components for graph preview (_graph.c), status bar (_status.c), and help overlay (_help.c). This completes the modularization of the milkCTRL overview dashboard rendering pipeline, improving code maintainability and separating concerns.
