+++ src/cli/CLIcore/CMakeLists.txt
@@ -225,6 +225,9 @@ add_executable(milkCTRL
     ../overview/overview_render_streams.c
     ../overview/overview_render_procs.c
     ../overview/overview_render_fps.c
+    ../overview/overview_render_graph.c
+    ../overview/overview_render_status.c
+    ../overview/overview_render_help.c
     ../overview/overview_input.c
     ../overview/overview_ctrl.c
     ../overview/stream_graph.c
