+++ src/cli/overview/overview_render.c
@@ -11,47 +11,7 @@
 
 #include "overview_render_internal.h"
 
-static double get_cpu_usage(void)
-{
-    static struct rusage last_usage;
-    static struct timespec last_time;
-    static int initialized = 0;
-    static double smoothed_cpu = 0.0;
-    
-    struct rusage current_usage;
-    struct timespec current_time;
-    
-    getrusage(RUSAGE_SELF, &current_usage);
-    clock_gettime(CLOCK_MONOTONIC, &current_time);
-    
-    if (!initialized) {
-        last_usage = current_usage;
-        last_time = current_time;
-        initialized = 1;
-        return 0.0;
-    }
-    
-    double dt = (current_time.tv_sec - last_time.tv_sec) +
-                (current_time.tv_nsec - last_time.tv_nsec) / 1e9;
-                
-    if (dt >= 0.5) { /* update every 0.5s */
-        double d_utime = (current_usage.ru_utime.tv_sec - last_usage.ru_utime.tv_sec) +
-                         (current_usage.ru_utime.tv_usec - last_usage.ru_utime.tv_usec) / 1e6;
-        double d_stime = (current_usage.ru_stime.tv_sec - last_usage.ru_stime.tv_sec) +
-                         (current_usage.ru_stime.tv_usec - last_usage.ru_stime.tv_usec) / 1e6;
-                         
-        double inst_cpu = 100.0 * (d_utime + d_stime) / dt;
-        smoothed_cpu = inst_cpu;
-        last_usage = current_usage;
-        last_time = current_time;
-    }
-    return smoothed_cpu;
-}
 
-/* All overview headers included via
- * overview_render_internal.h */
-
-/* forward declarations for scan API */
 extern float ov_scan_get_interval(void);
 
 /* =========================================================
@@ -561,6 +521,43 @@ void render_scroll_indicators(
     }
 }
 
+static double get_cpu_usage(void)
+{
+    static struct rusage last_usage;
+    static struct timespec last_time;
+    static int initialized = 0;
+    static double smoothed_cpu = 0.0;
+    
+    struct rusage current_usage;
+    struct timespec current_time;
+    
+    getrusage(RUSAGE_SELF, &current_usage);
+    clock_gettime(CLOCK_MONOTONIC, &current_time);
+    
+    if (!initialized) {
+        last_usage = current_usage;
+        last_time = current_time;
+        initialized = 1;
+        return 0.0;
+    }
+    
+    double dt = (current_time.tv_sec - last_time.tv_sec) +
+                (current_time.tv_nsec - last_time.tv_nsec) / 1e9;
+                
+    if (dt >= 0.5) { /* update every 0.5s */
+        double d_utime = (current_usage.ru_utime.tv_sec - last_usage.ru_utime.tv_sec) +
+                         (current_usage.ru_utime.tv_usec - last_usage.ru_utime.tv_usec) / 1e6;
+        double d_stime = (current_usage.ru_stime.tv_sec - last_usage.ru_stime.tv_sec) +
+                         (current_usage.ru_stime.tv_usec - last_usage.ru_stime.tv_usec) / 1e6;
+                         
+        double inst_cpu = 100.0 * (d_utime + d_stime) / dt;
+        smoothed_cpu = inst_cpu;
+        last_usage = current_usage;
+        last_time = current_time;
+    }
+    return smoothed_cpu;
+}
+
 void ov_render_header(
     OV_LAYOUT       *lay,
     const OV_MODEL  *m)
@@ -661,627 +658,6 @@ void ov_render_header(
     }
 
     ov_theme_bg(OV_BG_HEADER);
-}
-
-/**
- * ov_render_preview_line - full-width preview of
- *     the currently selected item (dashboard only).
- *
- * Renders on row 2, between header and panels.
- * Shows untruncated fields for the focused panel's
- * selected item.
- */
-static void ov_render_preview_line(
-    const OV_LAYOUT *lay,
-    const OV_MODEL  *m)
-{
-    int W = lay->term_cols;
-
-    ov_buf_pos(2, 1);
-    ov_theme_bg(OV_BG_PANEL);
-    ov_buf_hline(' ', W);
-    ov_buf_pos(2, 1);
-
-    /* Use frozen selection when freeze is active */
-    ov_focus_t focus = lay->freeze
-                       ? lay->freeze_focus
-                       : lay->focus;
-    int ssel = lay->freeze
-               ? lay->freeze_sel_stream
-               : lay->sel_stream;
-    int psel = lay->freeze
-               ? lay->freeze_sel_proc
-               : lay->sel_proc;
-    int fsel = lay->freeze
-               ? lay->freeze_sel_fps
-               : lay->sel_fps;
-
-    char line[512];
-    int  len = 0;
-    ov_rgb_t label_color = OV_FG_DIM;
-
-    switch (focus)
-    {
-    case OV_FOCUS_STREAMS:
-    {
-        label_color = OV_FG_STREAM;
-        if (ssel < 0 || ssel >= m->nb_streams)
-        {
-            break;
-        }
-        const OV_STREAM *s = &m->streams[ssel];
-        char szb[32];
-        if (s->naxis == 1)
-        {
-            snprintf(szb, sizeof(szb), "%u",
-                (unsigned) s->size[0]);
-        }
-        else if (s->naxis == 2)
-        {
-            snprintf(szb, sizeof(szb), "%ux%u",
-                (unsigned) s->size[0],
-                (unsigned) s->size[1]);
-        }
-        else
-        {
-            snprintf(szb, sizeof(szb),
-                "%ux%ux%u",
-                (unsigned) s->size[0],
-                (unsigned) s->size[1],
-                (unsigned) s->size[2]);
-        }
-        len = snprintf(line, sizeof(line),
-            " STM  %s  %s %s"
-            "  Hz:%.1f  ino:%lu"
-            "  own:%d  cnt:%lu"
-            "  wpid:%d  sem:%d",
-            s->name,
-            render_dtype(s->datatype),
-            szb,
-            s->update_hz,
-            (unsigned long) s->inode,
-            (int) s->ownerPID,
-            (unsigned long) s->cnt0,
-            (int) s->write_pid,
-            s->nb_sem);
-        break;
-    }
-    case OV_FOCUS_PROCS:
-    {
-        label_color = OV_FG_PROC;
-        if (psel < 0 || psel >= m->nb_procs)
-        {
-            break;
-        }
-        const OV_PROC *p = &m->procs[psel];
-        const char *sl;
-        switch (p->loopstat)
-        {
-        case 0:  sl = "IDLE"; break;
-        case 1:  sl = "RUN";  break;
-        case 2:  sl = "PAUS"; break;
-        case 3:  sl = "TERM"; break;
-        case 4:  sl = "ERR";  break;
-        default: sl = "??";   break;
-        }
-        len = snprintf(line, sizeof(line),
-            " PRC  %s  PID:%d  %s"
-            "  Hz:%.1f  trig:%s"
-            "  sem:%d  loop:%ld"
-            "  miss:%d  prio:%d",
-            p->name, (int) p->PID, sl,
-            p->loop_hz,
-            p->trigstreamname[0]
-                ? p->trigstreamname : "-",
-            p->triggersem,
-            (long) p->loopcnt,
-            p->triggermissed,
-            p->rt_priority);
-        break;
-    }
-    case OV_FOCUS_FPS:
-    {
-        label_color = OV_FG_FPS;
-        if (fsel < 0 || fsel >= m->nb_fps)
-        {
-            break;
-        }
-        const OV_FPS *f = &m->fps[fsel];
-        len = snprintf(line, sizeof(line),
-            " FPS  %s  C:%s R:%s"
-            "  st:%08X  cpid:%d  rpid:%d"
-            "  %s",
-            f->name,
-            f->conf_alive ? "Y" : "-",
-            f->run_alive  ? "Y" : "-",
-            f->md_status,
-            (int) f->confpid,
-            (int) f->runpid,
-            f->description);
-        break;
-    }
-    default:
-        break;
-    }
-
-    if (len > 0)
-    {
-        /* Label badge */
-        ov_theme_bg(label_color);
-        ov_buf_fg(0, 0, 0);
-        ov_buf_bold();
-        ov_buf_printf("%.*s", 5, line);
-        ov_buf_reset_attr();
-
-        /* Remaining content */
-        ov_theme_bg(OV_BG_PANEL);
-        ov_theme_fg(OV_FG_TEXT);
-        int rem = len - 5;
-        if (rem > W - 5)
-        {
-            rem = W - 5;
-        }
-        if (rem > 0)
-        {
-            ov_buf_printf("%.*s", rem, line + 5);
-        }
-    }
-
-    /* [FREEZE] badge on right edge when frozen */
-    if (lay->freeze)
-    {
-        const char *badge = " FREEZE ";
-        int bw = 8;
-        int col = W - bw + 1;
-        if (col > 1)
-        {
-            ov_buf_pos(2, col);
-            ov_buf_bg(60, 130, 200);
-            ov_buf_fg(255, 255, 255);
-            ov_buf_bold();
-            ov_buf_printf("%s", badge);
-        }
-    }
-
-    ov_buf_reset_attr();
-}
-
-
-void ov_render_graph_panel(
-    const OV_LAYOUT *lay,
-    const OV_MODEL  *m)
-{
-    OV_RECT r = lay->r_graph;
-    ov_draw_panel_border(r.row, r.col, r.height, r.width, "CONNECTIONS", OV_FG_CONN, lay->focus == OV_FOCUS_GRAPH);
-
-    int max_rows = r.height - 3;
-    int row = r.row + 1;
-
-    /* Render Header */
-    ov_buf_pos(row, r.col + 1);
-    ov_theme_bg(OV_BG_HEADER);
-    ov_theme_fg(OV_FG_DIM);
-    char htext[256];
-    int hlen = snprintf(htext, sizeof(htext), " %-12s %-4s %-12s %-6s %-16s %-4s %-5s %-4s", 
-                        "PROCESS", "PID", "STREAM", "STATUS", "LOOPCNT", "MISS", "MEM", "PR");
-    ov_buf_printf("%s", htext);
-    render_pad_spaces(hlen, r.width);
-    row++;
-
-    if (m->nb_edges == 0)
-    {
-        ov_buf_pos(row, r.col + 1);
-        ov_theme_bg(OV_BG_PANEL);
-        ov_buf_printf("  ");
-        ov_theme_fg(OV_FG_DIM);
-        ov_buf_printf("No connections detected");
-        render_pad_spaces(2 + 23, r.width);
-        row++;
-        for (int i = 1; i < max_rows; i++, row++) {
-            clear_row(row, r.col + 1, r.width - 2, OV_BG_PANEL);
-        }
-        ov_buf_reset_attr();
-        return;
-    }
-
-    int rendered_rows = 0;
-    for (int ei = 0; ei < m->nb_edges && rendered_rows < max_rows; ei++)
-    {
-        const OV_EDGE *e = &m->edges[ei];
-        if (e->src_node < 0 || e->src_node >= m->nb_nodes || e->tgt_node < 0 || e->tgt_node >= m->nb_nodes) continue;
-
-        const OV_NODE *src = &m->nodes[e->src_node];
-        const OV_NODE *tgt = &m->nodes[e->tgt_node];
-
-        ov_buf_pos(row, r.col + 1);
-        
-        int is_sel = (ei == lay->sel_graph && lay->focus == OV_FOCUS_GRAPH);
-        ov_rgb_t row_bg = is_sel ? OV_BG_SELECTED : OV_BG_PANEL;
-        ov_theme_bg(row_bg);
-        ov_buf_printf(" ");
-
-        /* Detailed rendering with color */
-        ov_rgb_t sc;
-        const char *styp = "UNK";
-        switch (src->type) {
-        case OV_NODE_STREAM: sc = OV_FG_STREAM; styp = "STRM"; break;
-        case OV_NODE_FPS:    sc = OV_FG_FPS;    styp = "FPS "; break;
-        case OV_NODE_PROC:   sc = OV_FG_PROC;   styp = "PROC"; break;
-        }
-
-        ov_rgb_t tc;
-        const char *ttyp = "UNK";
-        switch (tgt->type) {
-        case OV_NODE_STREAM: tc = OV_FG_STREAM; ttyp = "STRM"; break;
-        case OV_NODE_FPS:    tc = OV_FG_FPS;    ttyp = "FPS "; break;
-        case OV_NODE_PROC:   tc = OV_FG_PROC;   ttyp = "PROC"; break;
-        }
-
-        const char *etype = "UNKNOWN";
-        switch (e->type) {
-        case OV_EDGE_PROC_WRITES_STREAM:   etype = "PROC_WRITES_STRM"; break;
-        case OV_EDGE_STREAM_TRIGGERS_PROC: etype = "STRM_TRIGS_PROC "; break;
-        case OV_EDGE_FPS_RUNS_PROC:        etype = "FPS_RUNS_PROC   "; break;
-        case OV_EDGE_FPS_INPUT_STREAM:     etype = "FPS_INPUT_STRM  "; break;
-        case OV_EDGE_FPS_OUTPUT_STREAM:    etype = "FPS_OUTPUT_STRM "; break;
-        case OV_EDGE_PROC_TRIGGER_STREAM:  etype = "PROC_TRIGS_STRM "; break;
-        }
-
-        int printed = 1;
-        int avail = r.width - 2;
-
-        #define GRAPH_FIELD(color, fmt, ...)         \
-        do {                                         \
-            char _fb[128];                           \
-            int _fl = snprintf(                      \
-                _fb, sizeof(_fb), fmt,               \
-                ##__VA_ARGS__);                      \
-            int _vis = _fl;                          \
-            int _max = avail - printed;              \
-            if (_vis > _max) _vis = _max;            \
-            if (_vis > 0) {                          \
-                ov_theme_fg(color);                  \
-                ov_buf_printf("%.*s", _vis, _fb);    \
-                printed += _vis;                     \
-            }                                        \
-        } while(0)
-
-        GRAPH_FIELD(sc, "%-12.12s", src->name);
-        
-        /* Box drawing character length workaround */
-        if (avail - printed >= 4) {
-            ov_theme_fg(OV_FG_CONN);
-            ov_buf_printf(" %s%s ", OV_BOX_H, OV_TRI_R);
-            printed += 4;
-        }
-
-        GRAPH_FIELD(tc, "%-12.12s", tgt->name);
-        GRAPH_FIELD(OV_FG_DIM, " [%.6s]", e->label);
-        GRAPH_FIELD(OV_FG_TEXT, " %-16.16s ", etype);
-        GRAPH_FIELD(sc, "%-4s ", styp);
-        GRAPH_FIELD(tc, "%-4s", ttyp);
-
-        #undef GRAPH_FIELD
-
-        render_pad_spaces(printed, r.width);
-        row++;
-        rendered_rows++;
-    }
-    
-    for (; rendered_rows < max_rows; rendered_rows++, row++)
-    {
-        clear_row(row, r.col + 1, r.width - 2, OV_BG_PANEL);
-    }
-    
-    ov_buf_reset_attr();
-}
-
-void ov_render_status(
-    const OV_LAYOUT *lay,
-    const OV_MODEL  *m)
-{
-    OV_RECT r = lay->r_status;
-    ov_buf_pos(r.row, r.col);
-    ov_theme_bg(OV_BG_HEADER);
-    ov_theme_fg(OV_FG_DIM);
-
-    double interval = (double) ov_scan_get_interval();
-    double rate_hz  = (interval > 0.0) ? (1.0 / interval) : 0.0;
-
-    /* Context-sensitive ctrl hints */
-    const char *ctrl_hint = "";
-    if (lay->ctrl_mode)
-    {
-        switch (lay->focus)
-        {
-        case OV_FOCUS_FPS:
-            ctrl_hint = "  r:run s:conf";
-            break;
-        case OV_FOCUS_STREAMS:
-            ctrl_hint = "  d:delete";
-            break;
-        case OV_FOCUS_PROCS:
-            ctrl_hint = "  k:kill";
-            break;
-        default:
-            ctrl_hint = "";
-            break;
-        }
-    }
-
-    /* Sort key label */
-    const char *sort_label = "";
-    switch (lay->focus)
-    {
-    case OV_FOCUS_STREAMS:
-        switch (lay->sort_key_stream)
-        {
-        case 1:  sort_label = " [sort:typ]";   break;
-        case 2:  sort_label = " [sort:size]";  break;
-        case 3:  sort_label = " [sort:Hz]";    break;
-        case 4:  sort_label = " [sort:MB/s]";  break;
-        case 5:  sort_label = " [sort:inode]"; break;
-        case 6:  sort_label = " [sort:count]"; break;
-        default: sort_label = " [sort:name]";  break;
-        }
-        break;
-    case OV_FOCUS_PROCS:
-        switch (lay->sort_key_proc)
-        {
-        case 1:  sort_label = " [sort:PID]";  break;
-        case 2:  sort_label = " [sort:stat]"; break;
-        case 3:  sort_label = " [sort:Hz]";   break;
-        case 4:  sort_label = " [sort:MEM]";  break;
-        default: sort_label = " [sort:name]"; break;
-        }
-        break;
-    case OV_FOCUS_FPS:
-        switch (lay->sort_key_fps)
-        {
-        case 1:  sort_label = " [sort:alive]"; break;
-        case 2:  sort_label = " [sort:MEM]";   break;
-        default: sort_label = " [sort:name]";  break;
-        }
-        break;
-    default:
-        break;
-    }
-
-    const char *detail_label =
-        lay->detail_mode ? " [DETAIL]" : "";
-
-    /* Filter editing prompt overrides normal status */
-    if (lay->filter_editing)
-    {
-        const char *fstr = "";
-        switch (lay->focus)
-        {
-        case OV_FOCUS_STREAMS:
-            fstr = lay->filter_stream;
-            break;
-        case OV_FOCUS_PROCS:
-            fstr = lay->filter_proc;
-            break;
-        case OV_FOCUS_FPS:
-            fstr = lay->filter_fps;
-            break;
-        default:
-            break;
-        }
-        ov_theme_fg(OV_FG_TEXT);
-        int np = snprintf(
-            NULL, 0, " /%s", fstr);
-        ov_buf_printf(" /%s", fstr);
-
-        /* Blinking cursor */
-        ov_buf_fg(255, 200, 50);
-        ov_buf_printf("█");
-        np += 1;
-
-        ov_theme_fg(OV_FG_DIM);
-        ov_buf_printf("  (ENTER=accept ESC=cancel)");
-        np += 26;
-
-        int pad = r.width - np - 1;
-        if (pad > 0)
-        {
-            ov_buf_hline(' ', pad);
-        }
-        ov_buf_reset_attr();
-        return;
-    }
-
-    int n1 = 0;
-    if (lay->paused)
-    {
-        ov_buf_fg(255, 50, 50);
-        n1 = snprintf(NULL, 0, " [PAUSED] ");
-        ov_buf_printf(" [PAUSED] ");
-        ov_theme_fg(OV_FG_DIM);
-    }
-    else
-    {
-        n1 = snprintf(NULL, 0, " scan:%.0fms %.1fHz", m->scan_time_ms, rate_hz);
-        ov_buf_printf(" scan:%.0fms %.1fHz", m->scan_time_ms, rate_hz);
-    }
-
-    int n_hints = snprintf(NULL, 0,
-        "%s%s%s  +/- TAB D S/s / p c h q",
-        ctrl_hint, sort_label, detail_label);
-    ov_buf_printf(
-        "%s%s%s  +/- TAB D S/s / p c h q",
-        ctrl_hint, sort_label, detail_label);
-    n1 += n_hints;
-
-    time_t now = time(NULL);
-    struct tm *tm_ptr = localtime(&now);
-    char tstr[16];
-    int n2 = strftime(tstr, sizeof(tstr), "%H:%M:%S", tm_ptr);
-
-    /* Leave 1 char at the end to prevent terminal auto-scroll on last row */
-    int pad = r.width - n1 - n2 - 1;
-    if (pad > 0)
-    {
-        ov_buf_hline(' ', pad);
-    }
-    ov_theme_fg(OV_FG_TEXT);
-    ov_buf_printf("%s ", tstr);
-    ov_buf_reset_attr();
-}
-
-/* ov_render_help
- * Draws a centered help overlay.
- *
- * Algorithm:
- *   1. Enumerate all lines into a static table so pw/ph are exact.
- *   2. Fill the entire terminal with a dim translucent background
- *      BEFORE drawing the box — every terminal cell is written exactly
- *      once, eliminating the panel-under-help flicker.
- *   3. Draw border + content.
- */
-void ov_render_help(const OV_LAYOUT *lay)
-{
-    /* ---- Content table ---- */
-    static const struct
-    {
-        const char *text;
-        int         color_row; /* 1 = render as colour-legend row */
-    } lines[] =
-    {
-        { "Navigation",                              0 },
-        { "  F2-F6 / ^Left/^Right  Switch views",    0 },
-        { "  TAB      Cycle panel focus",            0 },
-        { "  UP/DOWN  Navigate list",                0 },
-        { "  Left/Right  Panel focus / scroll",      0 },
-        { "  PgUp/Dn  Scroll page",                  0 },
-        { "  Home/End  Jump to top/bottom",          0 },
-        { "Sorting",                                 0 },
-        { "  </>      Change sort column",           0 },
-        { "  [        Toggle sort direction",        0 },
-        { "  S        Sort by activity (Hz)",        0 },
-        { "  s        Sort by name",                 0 },
-        { "Display",                                 0 },
-        { "  +/-      Adjust scan rate",             0 },
-        { "  D        Toggle detail pane",           0 },
-        { "  L        Toggle lineage mode",          0 },
-        { "  p        Pause/resume display",         0 },
-        { "  SPACE    Freeze selection highlight",   0 },
-        { "  /        Filter (regex search)",        0 },
-        { "  W        Export snapshot to file",      0 },
-        { "  h        Toggle this help",             0 },
-        { "  q        Quit",                         0 },
-        { "Control mode  (c to toggle)",             0 },
-        { "  FPS:  r=run  s=conf",                   0 },
-        { "  STRM: d=delete stream",                 0 },
-        { "Process signals  (PROCS & FPS panels)",   0 },
-        { "  k  graceful kill (SIGTERM)",            0 },
-        { "  K  immediate kill (SIGKILL)",           0 },
-        { "  x  pause/resume (SIGSTOP/SIGCONT)",     0 },
-        { "Detail View (ENTER or D)",                0 },
-        { "  Toggles detail pane for selected item", 0 },
-        { "Columns",                                 0 },
-        { "  STRM: MB/s throughput, total in footer",0 },
-        { "  PROC: DUTY% (exec/iter), CPU%, MEM",    0 },
-        { "  FPS:  MEM (RSS)",                       0 },
-        { "Mouse",                                     0 },
-        { "  Click=select  DblClick=detail",           0 },
-        { "  Scroll wheel=navigate list",              0 },
-        { "",                                          0 },
-        { "Colors:  ",                                 1 },
-    };
-    static const int NL = (int)(sizeof(lines) / sizeof(lines[0]));
-
-    /* Compute exact box dimensions from content */
-    int content_w = 0;
-    for (int i = 0; i < NL; i++)
-    {
-        int l = (int) strlen(lines[i].text);
-        if (l > content_w)
-        {
-            content_w = l;
-        }
-    }
-    /* Add " stream proc fps" for the color legend row */
-    int legend_extra = (int) strlen(" stream proc fps");
-    int last_text_w  =
-        (int) strlen(lines[NL - 1].text) + legend_extra;
-    if (last_text_w > content_w)
-    {
-        content_w = last_text_w;
-    }
-
-    /* Box: 2-char left pad + content + 2-char right pad + 2 borders */
-    int pw = content_w + 4 + 2;
-    /* Box height: 2 borders + 1 top margin + NL lines + 1 bottom margin */
-    int ph = NL + 4;
-
-    int W  = lay->term_cols;
-    int H  = lay->term_rows;
-
-    int pr = (H - ph) / 2;
-    int pc = (W - pw) / 2;
-    if (pr < 1) { pr = 1; }
-    if (pc < 1) { pc = 1; }
-
-    /* ---- Step 1: dim overlay for the whole terminal ---- */
-    /* REMOVED: no longer overwriting the whole terminal so background
-     * panels remain visible. */
-
-    /* ---- Step 2: border + interior ---- */
-    ov_draw_panel_border(pr, pc, ph, pw, "HELP", OV_FG_BRIGHT, 1);
-
-    for (int r = pr + 1; r < pr + ph - 1; r++)
-    {
-        clear_row(r, pc + 1, pw - 2, OV_BG_PANEL);
-    }
-
-    /* ---- Step 3: text content ---- */
-    int row = pr + 2;
-    for (int i = 0; i < NL; i++)
-    {
-        ov_buf_pos(row, pc + 3);
-        ov_theme_bg(OV_BG_PANEL);
-
-        /* Section headers: non-empty lines that
-         * don't start with a space */
-        const char *t = lines[i].text;
-        int is_section = (t[0] != '\0'
-                          && t[0] != ' '
-                          && !lines[i].color_row);
-        if (is_section)
-        {
-            ov_theme_fg(OV_FG_TITLE);
-            ov_buf_bold();
-        }
-        else
-        {
-            ov_theme_fg(OV_FG_TEXT);
-            ov_buf_reset_attr();
-            ov_theme_bg(OV_BG_PANEL);
-        }
-
-        if (lines[i].color_row)
-        {
-            /* Color legend: print label then colored words */
-            ov_theme_fg(OV_FG_DIM);
-            ov_buf_printf("%s", lines[i].text);
-            ov_theme_fg(OV_FG_STREAM);
-            ov_buf_printf("stream ");
-            ov_theme_fg(OV_FG_PROC);
-            ov_buf_printf("proc ");
-            ov_theme_fg(OV_FG_FPS);
-            ov_buf_printf("fps");
-        }
-        else
-        {
-            ov_buf_printf("%-*s", content_w, lines[i].text);
-        }
-
-        ov_buf_reset_attr();
-        row++;
-    }
-
-    ov_buf_reset_attr();
 }
 
 void ov_render_frame(
