+++ src/cli/overview/overview_render_internal.h
@@ -181,4 +181,18 @@ int ov_render_detail_panel(
     const OV_LAYOUT *lay,
     const OV_MODEL  *m);
 
+void ov_render_graph_panel(
+    const OV_LAYOUT *lay,
+    const OV_MODEL  *m);
+
+void ov_render_status(
+    const OV_LAYOUT *lay,
+    const OV_MODEL  *m);
+
+
+extern float ov_scan_get_interval(void);
+
+
 #endif /* OVERVIEW_RENDER_INTERNAL_H */
+void ov_render_help(const OV_LAYOUT *lay);
+void ov_render_preview_line(const OV_LAYOUT *lay, const OV_MODEL *m);
